RCPTT launch updates #145
Removes redundant configuration of macOS-specific start on first thread
feature; updated the way vm arguments are set to the application under
test. Furthermore, hoping the <vm> parameter will ensure that the
application under test will start on a toolchain-aware Java 11 instance
(on the server it starts with Java 17).

The hope is that by reducing the maximum amount of memory used by the
RCPTT build, it becomes less often that the build environment runs out
of memory (that might be the possible cause of random test failures).