Fixes normalization error in duplicate checker #178
When updating to Xtext 2.35, the duplicate pattern tests failed because
in the JUnit settings a different uri normalization approach was used
(but strangely not when running the tests from Eclipse). This test
ensures that the same approach is used to verify duplicates in the index
that is used by the Xtext index itself to avoid such issues in the
future.